chore: :fire: remove unused ClockInterface and DefaultClaimsFactory
Eliminate ClockInterface and DefaultClaimsFactory to simplify the codebase and reduce dependencies.
